requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-filter==21.1
django-import-export==2.7.1
django-model-utils==4.2.0
django-smart-selects==1.5.9
djangorestframework==3.13.1
et-xmlfile==1.1.0
install==1.3.5
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pyasn1==0.4.8
python3-ldap==0.9.8.4
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
unicodecsv==0.14.1
xlrd==2.0.1
xlwt==1.3.0
